RPCT-2312, Monitor: Added static code analyzer to Monitor project

+++ Monitor/MonitorMainWindow.cpp
@@ -216,7 +216,8 @@ void MonitorMainWindow::restoreWindowState()
 		(frameGeometry().width() > screenRect.width() ||
 		frameGeometry().height() > screenRect.height()))
 	{
-		resize(screenRect.width() * 0.7, screenRect.height() * 0.7);
+		resize(static_cast<int>(screenRect.width() * 0.7),
+			   static_cast<int>(screenRect.height() * 0.7));
 	}
 
 	return;
